backend/requirements.txt
Django==5.2.5
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.7.0
python-dotenv==1.1.1
openai==1.101.0
requests==2.32.5
drf-spectacular==0.27.2
Pillow==10.4.0
pytz==2023.3
gunicorn==22.0.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
